+++ CHANGELOG.md
@@ -32,6 +32,7 @@
     * Fix issue where tests using the check package cannot be run using the `Go: Test File` command. [Bug 1911](https://github.com/Microsoft/vscode-go/issues/1911)
     * Disable module support when installing the Go tools. Fixes [Bug 1919](https://github.com/Microsoft/vscode-go/issues/1919)
     * Use version 2 of delve apis by default instead of version 1. [Feature Request 1876](https://github.com/Microsoft/vscode-go/issues/1876)
+    * Prompt to update `gogetdoc` and `gopkgs` if using Go modules.
     * Add support for the below features when using Go modules
         * `Go: Add Import` command that gives you a list of importable packages to add to the import block of current file
         * `Go: Browse Packages` command that lets you browse available packages and their files